+++ R/tbl_sum.R
@@ -12,10 +12,12 @@
 #'   and the data source in the name of the first element.
 #'
 #' @seealso [pillar::type_sum()]
-#' @param x Object to summarise
+#' @param x Object to summarise.
 #' @export
 # Can be overridden in .onLoad()
-tbl_sum <- function(x) UseMethod("tbl_sum", x)
+tbl_sum <- function(x) {
+  UseMethod("tbl_sum", x)
+}
 
 # If needed, exported in .onLoad() via replace_if_pillar_has()
 tbl_sum.default <- function(x) {
@@ -24,6 +26,11 @@ tbl_sum.default <- function(x) {
 
 # If needed, exported in .onLoad() via replace_if_pillar_has()
 tbl_sum.tbl <- function(x) {
+  c("A data frame" = dim_desc(x))
+}
+
+# Must be registered in .onLoad() until pillar 1.5.0 is on CRAN
+tbl_sum.tbl_df <- function(x) {
   c("A tibble" = dim_desc(x))
 }