+++ R/zzz.R
@@ -1,3 +1,4 @@
+# nocov start
 .onLoad <- function(libname, pkgname) {
   if (getRversion() == "3.3.0") {
     safe_match <<- safe_match_3_0
@@ -5,22 +6,22 @@
     safe_match <<- safe_match_default
   }
 
+  # Simplify once pillar 1.5.0 is on CRAN
   replace_if_pillar_has(glimpse)
-  register_if_pillar_hasnt(glimpse.default, glimpse.data.frame, glimpse.tbl)
   replace_if_pillar_has(tbl_sum)
-  register_if_pillar_hasnt(tbl_sum.default, tbl_sum.tbl)
-  register_if_pillar_hasnt(format.tbl)
-  register_if_pillar_hasnt(print.tbl)
+  register_if_pillar_hasnt(
+    glimpse.default, glimpse.data.frame, glimpse.tbl,
+    tbl_sum.default, tbl_sum.tbl, format.tbl, print.tbl
+  )
+
+  vctrs::s3_register("tibble::tbl_sum", "tbl_df")
 }
 
 safe_match_3_0 <- function(x, table) {
-  # nocov start
   match(x, table, incomparables = character())
-  # nocov end
 }
 
 safe_match_default <- function(x, table) {
-  # nocov start
   match(x, table)
-  # nocov end
 }
+# nocov end